Back to one test file for ingestionManager + update count
