+++ opencti-platform/opencti-graphql/tests/utils/syncCountHelper.ts
@@ -39,9 +39,9 @@ testCreatedCounter.infrastructure = 1;
 testCreatedCounter['intrusion-set'] = 4;
 testCreatedCounter['ipv4-addr'] = 1;
 testCreatedCounter['kill-chain-phase'] = 3;
-testCreatedCounter.label = 19;
+testCreatedCounter.label = 15;
 testCreatedCounter.language = 1;
-testCreatedCounter.location = 24;
+testCreatedCounter.location = 21;
 testCreatedCounter['mac-addr'] = 1;
 testCreatedCounter.malware = 50;
 testCreatedCounter['malware-analysis'] = 3;
@@ -73,7 +73,7 @@ testUpdatedCounter.malware = 19;
 testUpdatedCounter.file = 19;
 testUpdatedCounter['intrusion-set'] = 4;
 testUpdatedCounter['data-component'] = 7;
-testUpdatedCounter.location = 15;
+testUpdatedCounter.location = 14;
 testUpdatedCounter['attack-pattern'] = 3;
 testUpdatedCounter['case-incident'] = 11;
 testUpdatedCounter.feedback = 1;
